requirements.txt
flask
pandas
numpy
python-dotenv
requests
googlemaps
ortools
scikit-learn
folium
flask-login

